+++ integration-tests/tests/src/tests/sync/flexible.ts
@@ -29,12 +29,12 @@
 // fraction too long.
 
 import { expect } from "chai";
-import Realm, { BSON, ClientResetMode, SessionStopPolicy } from "realm";
+import Realm, { BSON, ClientResetMode, FlexibleSyncConfiguration, SessionStopPolicy } from "realm";
 
 import { authenticateUserBefore, importAppBefore, openRealmBeforeEach } from "../../hooks";
 import { DogSchema, IPerson, PersonSchema } from "../../schemas/person-and-dog-with-object-ids";
 import { expectClientResetError } from "../../utils/expect-sync-error";
-import { closeAndReopenRealm } from "../../utils/close-realm";
+import { closeAndReopenRealm, closeRealm, closeThisRealm } from "../../utils/close-realm";
 
 const FlexiblePersonSchema = { ...PersonSchema, properties: { ...PersonSchema.properties, nonQueryable: "string?" } };
 
@@ -154,8 +154,8 @@ describe.skipIf(environment.missingServer, "Flexible sync", function () {
 
       it("does not accept { flexible: true } and a partition value", function () {
         expect(() => {
-          // Cast to any as Typescript will detect this as an error
-          new (Realm as any)({
+          // @ts-expect-error Intentionally testing the wrong type
+          new Realm({
             sync: {
               _sessionStopPolicy: SessionStopPolicy.Immediately,
               flexible: true,
@@ -195,8 +195,8 @@ describe.skipIf(environment.missingServer, "Flexible sync", function () {
 
       it("throws an error if flexible sync is enabled and client reset mode is discardLocal", function () {
         expect(() => {
-          // Cast to any as Typescript will detect this as an error
-          new (Realm as any)({
+          // @ts-expect-error Intentionally testing the wrong type
+          new Realm({
             sync: {
               _sessionStopPolicy: SessionStopPolicy.Immediately,
               flexible: true,
@@ -207,6 +207,179 @@ describe.skipIf(environment.missingServer, "Flexible sync", function () {
             },
           });
         }).to.throw("Only manual client resets are supported with flexible sync");
+      });
+
+      describe("initialSubscriptions option", function () {
+        function getConfig(
+          user: Realm.User,
+          initialSubscriptions: Realm.FlexibleSyncConfiguration["initialSubscriptions"],
+        ): Realm.Configuration {
+          return {
+            schema: [FlexiblePersonSchema, DogSchema],
+            sync: {
+              _sessionStopPolicy: SessionStopPolicy.Immediately,
+              flexible: true,
+              user,
+              initialSubscriptions,
+            },
+          };
+        }
+
+        describe("error", function () {
+          afterEach(function () {
+            Realm.deleteFile(this.config);
+          });
+
+          it("throws an error if no update function is provided", async function (this: RealmContext) {
+            // @ts-expect-error Intentionally testing the wrong type
+            this.config = getConfig(this.user, {});
+
+            await expect(Realm.open(this.config)).to.be.rejectedWith(
+              "update must be of type 'function', got (undefined)",
+            );
+          });
+
+          it("throws an error if update is undefined", async function (this: RealmContext) {
+            this.config = getConfig(this.user, {
+              // @ts-expect-error Intentionally testing the wrong type
+              update: undefined,
+            });
+
+            await expect(Realm.open(this.config)).to.be.rejectedWith(
+              "update must be of type 'function', got (undefined)",
+            );
+          });
+
+          it("throws an error if update is not a function", async function (this: RealmContext) {
+            this.config = getConfig(this.user, {
+              // @ts-expect-error Intentionally testing the wrong type
+              update: "Person",
+            });
+
+            await expect(Realm.open(this.config)).to.be.rejectedWith("update must be of type 'function', got (Person)");
+          });
+
+          it("throws an error if `rerunOnOpen` is not a boolean", async function (this: RealmContext) {
+            this.config = getConfig(this.user, {
+              update: () => {
+                // no-op
+              },
+              // @ts-expect-error Intentionally testing the wrong type
+              rerunOnOpen: "yes please",
+            });
+
+            await expect(Realm.open(this.config)).to.be.rejectedWith(/rerunOnOpen must be of type 'boolean', got.*/);
+          });
+        });
+
+        describe("success", function () {
+          type ExtraConfig = Partial<FlexibleSyncConfiguration["initialSubscriptions"]>;
+
+          function getSuccessConfig(user: Realm.User, extraConfig: ExtraConfig = {}) {
+            return getConfig(user, {
+              update: (subs, realm) => {
+                subs.add(realm.objects(FlexiblePersonSchema.name));
+              },
+              ...extraConfig,
+            });
+          }
+
+          async function openRealm(user: Realm.User, extraConfig: ExtraConfig = {}) {
+            const config = getSuccessConfig(user, extraConfig);
+            const realm = await Realm.open(config);
+
+            return { realm, config };
+          }
+
+          async function testSuccess(
+            user: Realm.User,
+            extraConfig: Partial<FlexibleSyncConfiguration["initialSubscriptions"]> = {},
+            closeRealmAfter = true,
+          ) {
+            const { realm, config } = await openRealm(user, extraConfig);
+
+            try {
+              expect(realm.subscriptions).to.have.length(1);
+              expect(realm.subscriptions.state).to.equal(Realm.App.Sync.SubscriptionsState.Complete);
+            } finally {
+              if (closeRealmAfter) {
+                closeRealm(realm, config);
+              }
+            }
+
+            return closeRealmAfter ? undefined : realm;
+          }
+
+          it("returns a promise", async function (this: RealmContext) {
+            const result = openRealm(this.user, {});
+            try {
+              expect(result).to.be.instanceOf(Promise);
+            } finally {
+              const { realm, config } = await result;
+              closeRealm(realm, config);
+            }
+          });
+
+          it("can be used with the `new Realm` constructor", async function (this: RealmContext) {
+            const config = getSuccessConfig(this.user);
+            const realm = new Realm(config);
+
+            await realm.subscriptions.waitForSynchronization();
+
+            expect(realm.subscriptions).to.have.length(1);
+            expect(realm.subscriptions.state).to.equal(Realm.App.Sync.SubscriptionsState.Complete);
+
+            closeRealm(realm, config);
+          });
+
+          it("updates the subscriptions on first open if rerunOnOpen is undefined", async function (this: RealmContext) {
+            await testSuccess(this.user);
+          });
+
+          it("updates the subscriptions on first open if rerunOnOpen is false", async function (this: RealmContext) {
+            await testSuccess(this.user, { rerunOnOpen: false });
+          });
+
+          it("updates the subscriptions on first open if rerunOnOpen is true", async function (this: RealmContext) {
+            await testSuccess(this.user, { rerunOnOpen: true });
+          });
+
+          it("does not update the subscriptions on second open if rerunOnOpen is undefined", async function (this: RealmContext) {
+            const realm = await testSuccess(this.user, {}, false);
+            if (!realm) throw new Error("Valid realm was not returned from testSuccess");
+
+            await realm.subscriptions.update((subs) => subs.removeAll());
+            realm.close();
+
+            const { realm: realm2, config } = await openRealm(this.user, {});
+            expect(realm2.subscriptions).to.have.length(0);
+
+            closeRealm(realm2, config);
+          });
+
+          it("does not update the subscriptions on second open if rerunOnOpen is false", async function (this: RealmContext) {
+            const realm = await testSuccess(this.user, { rerunOnOpen: false }, false);
+            if (!realm) throw new Error("Valid realm was not returned from testSuccess");
+
+            await realm.subscriptions.update((subs) => subs.removeAll());
+            realm.close();
+
+            const { realm: realm2, config } = await openRealm(this.user, {});
+            expect(realm2.subscriptions).to.have.length(0);
+
+            closeRealm(realm2, config);
+          });
+
+          it("does update the subscriptions on second open if rerunOnOpen is true", async function (this: RealmContext) {
+            const realm = await testSuccess(this.user, { rerunOnOpen: true }, false);
+            if (!realm) throw new Error("Valid realm was not returned from testSuccess");
+
+            await realm.subscriptions.update((subs) => subs.removeAll());
+            realm.close();
+
+            await testSuccess(this.user, { rerunOnOpen: true });
+          });
+        });
       });
     });
   });
@@ -235,7 +408,10 @@ describe.skipIf(environment.missingServer, "Flexible sync", function () {
       });
 
       it("throws an error if the Realm has a partition based sync config", function (this: RealmContext) {
-        const realm = new Realm({ schema: [FlexiblePersonSchema], sync: { user: this.user, partitionValue: "test" } });
+        const realm = new Realm({
+          schema: [FlexiblePersonSchema],
+          sync: { user: this.user, partitionValue: "test" },
+        });
         expect(() => realm.subscriptions).to.throw(
           "`subscriptions` can only be accessed if flexible sync is enabled, but partition based sync is currently enabled for your Realm. Modify your sync config to remove any `partitionValue` and enable flexible sync, for example: { sync: { user, flexible: true } }",
         );
@@ -1396,7 +1572,11 @@ describe.skipIf(environment.missingServer, "Flexible sync", function () {
       it("triggers a client reset if items are added without a subscription", async function (this: RealmContext) {
         await expectClientResetError(this.config, this.user, (realm) => {
           realm.write(() => {
-            return realm.create<IPerson>(FlexiblePersonSchema.name, { _id: new BSON.ObjectId(), name: "Tom", age: 36 });
+            return realm.create<IPerson>(FlexiblePersonSchema.name, {
+              _id: new BSON.ObjectId(),
+              name: "Tom",
+              age: 36,
+            });
           });
         });
       });
@@ -1409,7 +1589,11 @@ describe.skipIf(environment.missingServer, "Flexible sync", function () {
           this.user,
           (realm) => {
             realm.write(() => {
-              realm.create<IPerson>(FlexiblePersonSchema.name, { _id: new BSON.ObjectId(), name: "Tom Old", age: 99 });
+              realm.create<IPerson>(FlexiblePersonSchema.name, {
+                _id: new BSON.ObjectId(),
+                name: "Tom Old",
+                age: 99,
+              });
             });
           },
           () => {
@@ -1429,7 +1613,11 @@ describe.skipIf(environment.missingServer, "Flexible sync", function () {
           // Deliberately not waiting for synchronisation here
 
           realm.write(function () {
-            return realm.create<IPerson>(FlexiblePersonSchema.name, { _id: new BSON.ObjectId(), name: "Tom", age: 36 });
+            return realm.create<IPerson>(FlexiblePersonSchema.name, {
+              _id: new BSON.ObjectId(),
+              name: "Tom",
+              age: 36,
+            });
           });
         });
       });
@@ -1441,7 +1629,11 @@ describe.skipIf(environment.missingServer, "Flexible sync", function () {
           await addSubscriptionAndSync(realm, realm.objects(FlexiblePersonSchema.name).filtered("age < 40"));
 
           const person = realm.write(() => {
-            return realm.create<IPerson>(FlexiblePersonSchema.name, { _id: new BSON.ObjectId(), name: "Tom", age: 36 });
+            return realm.create<IPerson>(FlexiblePersonSchema.name, {
+              _id: new BSON.ObjectId(),
+              name: "Tom",
+              age: 36,
+            });
           });
 
           realm.write(() => {