+++ CHANGELOG.md
@@ -1,7 +1,31 @@
 x.x.x Release notes (yyyy-MM-dd)
 =============================================================
 ### Enhancements
-* None.
+* Added an `initialSubscriptions` option to the `sync` config, which allows users to specify a subscription update function to bootstrap a set of flexible sync subscriptions when the Realm is first opened (or every time the app runs, using the `rerunOnOpen` flag). (#4561[https://github.com/realm/realm-js/pull/4561])
+
+    Example usage:
+    ```ts
+    const thirtyDaysAgo = new Date(new Date().setDate(new Date().getDate() - 30));
+    const config = {
+      // ...
+      sync: {
+        flexible: true,
+        user,
+        initialSubscriptions: {
+          update: (realm) => {
+            realm.subscriptions.update((subs) => {
+              subs.add(
+                realm.objects("Person").filtered("dateOfBirth > $0", thirtyDaysAgo),
+                // This is a named subscription, so will replace any existing subscription with the same name
+                { name: "People30Days" }
+              );
+            });
+          },
+          rerunOnStartup: true,
+        },
+      },
+    };
+    ```
 
 ### Fixed
 * Flexible sync would not correctly resume syncing if a bootstrap was interrupted. ([realm/realm-core#5466](https://github.com/realm/realm-core/pull/5466), since v10.12.0)